Course Manager Status Set
The course manager can now set the status from his home page as well as
from the course pages he manages. Also corrected the moodle_url issues
as well as set  the after form submission action to redirect it to the
page it was called from using the get_referer() function. Also changed
the function header() to moodle function redirect().